+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add more hot reload `BuildArgs` helpers.
 * Change default hot reload rebuilder to first try env var `"ZNG_HOT_RELOAD_REBUILDER"`.
+* Fix `tracing` in hot reloaded dylib.
 
 # 0.6.0